requirements.txt
Django==3.2.5
djangorestframework==3.12.4
django-rest-auth==0.9.5
djangorestframework-simplejwt==4.7.1
django-cors-headers==3.7.0
